+++ www/server/watch.php
@@ -1,11 +1,28 @@
 <?php
 
+
+// options http://localhost/server/watch.php?dir=storage (wav)
+// http://localhost/server/watch.php?dir=inspect (mp3)
+
 include_once('config.inc.php');
 
-$dir = 'reception';
+$dir = getGET('dir', 'reception');
+
+$route = array(
+    'reception' => RECEPTION,
+    'inspect' => INSPECT,
+    'storage' => STORAGE
+);
+
+if(array_key_exists($dir, $route)) {
+    $scandir = $route[$dir];
+} else {
+    $scandir = RECEPTION;
+}
 $list = array();
 
-if ($handle = opendir(RECEPTION)) {
+
+if ($handle = opendir($scandir)) {
     while (false !== ($entry = readdir($handle))) {
         if(substr($entry, 0, 1) !== "."  && $entry != '.' && $entry != '..'){
             $list[] = $entry;